x86: adjust cr3_pcid() return type
There's no need for it to be 64 bits wide - only the low twelve bits
of CR3 hold the PCID.

Acked-by: Andrew Cooper <andrew.cooper3@citrix.com>

+++ xen/include/asm-x86/processor.h
@@ -287,7 +287,7 @@ static inline unsigned long cr3_pa(unsigned long cr3)
     return cr3 & X86_CR3_ADDR_MASK;
 }
 
-static inline unsigned long cr3_pcid(unsigned long cr3)
+static inline unsigned int cr3_pcid(unsigned long cr3)
 {
     return cr3 & X86_CR3_PCID_MASK;
 }
